data type change and notification fix

+++ db/migrate/20161229215010_add_column_to_cards.rb
@@ -1,6 +1,6 @@
 class AddColumnToCards < ActiveRecord::Migration
   def change
-    add_column :cards, :standard, :decimal
-    add_column :cards, :arena, :decimal
+    add_column :cards, :standard, :float
+    add_column :cards, :arena, :float
   end
 end
